requirements.txt
alembic==1.15.2
colorama==0.4.6
greenlet==3.2.1
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
